requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
django==3.2
    # via -r requirements.in
python-slugify==6.1.2
    # via -r requirements.in
pytz==2022.2.1
    # via django
sqlparse==0.4.2
    # via django
text-unidecode==1.3
    # via python-slugify
